Introduction to algorithms chapter 1
Solving problem 1-1
I'm hitting Go's max float64 limit for 14 cases.
Maybe an unsigned int 64 might reduce this, but I've got a feeling this
won't be enough and I'll have to resort to math/big.Float or something.

At first, I naively tried to call f(n) for 1 -> ONE KAZILLION, but after
a few solutions, the program started running for hours without yielding
a result :-)

Then by looking at f(n) -> n and trying to optimize that, it hit me that
I just needed to invert the given functions and use the different
timings as input for that.